moved banner into shared
