+++ packages/modules/src/modules/shared/ticker/Ticker.tsx
